+++ models/common.go
@@ -7,7 +7,7 @@ import (
 )
 
 type CommonModel struct {
-	ID        uint64         `gorm:"primaryKey" json:"id"`
+	ID        uint16         `gorm:"primaryKey" json:"id"`
 	CreatedAt time.Time      `gorm:"<-:create" json:"-"`
 	UpdatedAt time.Time      `json:"-"`
 	DeletedAt gorm.DeletedAt `gorm:"index" json:"-"`
